Resolves #21, cleanup javascript
Though the problem on ie9 could not be reproduced there was an error
in the javascript on the landingpage. $ was not yet loaded when the
script was loaded which gave an js error. Also the success function
of the ajax call was never called because of a parse error. The type
was text/html, so it tried to parse the empt response as html which
failed.